chore(backend): move token directory from modules to libs/modules ss-18
